+++ lib/core.dart
@@ -1,6 +1,4 @@
 export 'package:dart_corelib/dart_corelib.dart';
-export 'network/models/string_or.dart';
-export 'network/models/unix_time.dart';
 export 'system/converters/enum_converter.dart';
 export 'system/models/date_time_list.dart';
 export 'system/models/distance.dart';