+++ lib/network/consts/http_method.dart
@@ -1,7 +1,0 @@
-abstract class HttpMethod {
-  static const String get = 'GET';
-  static const String post = 'POST';
-  static const String delete = 'DELETE';
-  static const String patch = 'PATCH';
-  static const String put = 'PUT';
-}