+++ lib/network/dio_ext/dio_log_intercepter.dart
@@ -1,4 +1,0 @@
-@Deprecated('Import dio_log_interceptor.dart (correct spelling).')
-library;
-
-export 'dio_log_interceptor.dart';