+++ lib/network/models/result.dart
@@ -1,1 +1,0 @@
-export 'package:dart_corelib/models/result.dart';