+++ lib/network/utils/firebase_util.dart
@@ -1,5 +1,0 @@
-class FirebaseUtil {
-  static String parseErrorMessage(String errorMessage) {
-    return errorMessage.substring(errorMessage.indexOf(']') + 1);
-  }
-}