+++ lib/storage.dart
@@ -1,6 +1,7 @@
 export 'package:dart_corelib/cache/cache.dart';
 export 'package:dart_corelib/cache/cache_list.dart';
 export 'package:dart_corelib/cache/cache_map.dart';
+export 'prefs/auth_session.dart';
 export 'prefs/flutter_prefs.dart';
 export 'prefs/prefs.dart';
 export 'prefs/prefs_list.dart';